clear the whole vert_pred buffer for 24 bit decoding. This is currently not needed since the decoder creates only half of the pixels, but should reduce confusion for people fiddling with it *g*
Originally committed as revision 4687 to svn://svn.ffmpeg.org/ffmpeg/trunk

+++ libavcodec/truemotion1.c
@@ -744,7 +744,7 @@ static void truemotion1_decode_24bit(TrueMotion1Context *s)
     int index;
 
     /* clean out the line buffer */
-    memset(s->vert_pred, 0, s->avctx->width * sizeof(unsigned short));
+    memset(s->vert_pred, 0, s->avctx->width * sizeof(unsigned int));
 
     GET_NEXT_INDEX();
